+++ Wrappers/Python/ccpi/viewer/utils/conversion.py
@@ -1855,8 +1855,8 @@ class vtkImageResampler(VTKPythonAlgorithmBase):
             # (xy the same, z possibly different)
             if not self.GetIsAcquisitionData():
                 xy_axes_magnification = np.power(max_size / total_size, 1 / 3)
-                num_slices_per_chunk = int(
-                    1 / xy_axes_magnification)  # number of slices in the z direction we are resampling together.
+                num_slices_per_chunk = int(np.round(
+                    1 / xy_axes_magnification))  # number of slices in the z direction we are resampling together.
             else:
                 # If we have acquisition data we don't want to resample in the z
                 # direction because then we would be averaging projections together,
