+++ Wrappers/Python/test/test_io.py
@@ -8,17 +8,7 @@ from ccpi.viewer.utils.conversion import Converter
 from ccpi.viewer.utils.io import ImageReader, cilviewerHDF5Writer, cilviewerHDF5Reader
 
 
-def calculate_target_downsample_shape(max_size, total_size, shape, acq=False):
-    if not acq:
-        xy_axes_magnification = np.power(max_size / total_size, 1 / 3)
-        slice_per_chunk = int(1 / xy_axes_magnification)
-    else:
-        slice_per_chunk = 1
-        xy_axes_magnification = np.power(max_size / total_size, 1 / 2)
-    num_chunks = 1 + len([i for i in range(slice_per_chunk, shape[2], slice_per_chunk)])
-
-    target_image_shape = (int(xy_axes_magnification * shape[0]), int(xy_axes_magnification * shape[1]), num_chunks)
-    return target_image_shape
+from utils import calculate_target_downsample_shape
 
 
 class TestImageReaderAndWriter(unittest.TestCase):
@@ -85,7 +75,9 @@ class TestImageReaderAndWriter(unittest.TestCase):
     def _test_resampling_not_acq_data(self, reader, target_size):
         image = reader.Read()
         extent = image.GetExtent()
-        resulting_shape = (extent[1] + 1, (extent[3] + 1), (extent[5] + 1))
+        # resulting_shape = (extent[1] + 1, (extent[3] + 1), (extent[5] + 1))
+        resulting_shape = tuple(image.GetDimensions())
+        print (resulting_shape)
         og_shape = np.shape(self.input_3D_array)
         og_shape = (og_shape[2], og_shape[1], og_shape[0])
         og_size = og_shape[0] * og_shape[1] * og_shape[2] * self.bytes_per_element
